Remove AnyEvent from proxy process
This is to avoid problems with fork (documented in AnyEvent POD).
Basically if we fork and AnyEvent is used in the parent process it
prevents us to use it reliably in the child process (according to the
doc). I guess, we *might* be able to clean AE internals in the child,
enough to have a working child... but rather than taking the risk I've
converted proxy to a classic select() loop and SIGCHLD handling.