Adds 'auto' proxies, automatically starting/stopping
This is the default, unless 'noauto' is specified in the proxy settings.

Note that this doesn't completely satisfy me yet, because it might not
be completely obvious that the code for your process is not reloaded
on stop/start if there are other processes attached to the same proxy.

fixes gh-5